Use WaveTrackCache for spectrograms
Zoomin-in of spectrograms (or other view changes that invalidate the
whole pixel cache of the WaveClip) used to do at least one opening and
closing of a block file for each column of pixels.  With this change,
open each block file not more than once for each repopulation of the
cache.  It makes zooming noticeably faster on my feeble laptop.