Reorganize ProjectStatus, ProjectWindowBase, and AudacityProject...
... ProjectStatus and ProjectWindowBase cooperate closely and are fused into
one file.

The services added to wxFrame by ProjectWindowBase are the handling of status
bar messages, and event notifications when they change, and association with
a project object.

But that object is only a base of AudacityProject.  This new base class just
hold the pointers to its ProjectWindowBase, and to TrackPanel (as wxWindow
only).

These changes allow AButton and ASlider to avoid dependence on AudacityProject,
letting those widgets live in a lower layer of the dependency graph.